+++ include/seqan3/core/parallel/detail/all.hpp
@@ -13,5 +13,5 @@
 #pragma once
 
 #include <seqan3/core/parallel/detail/latch.hpp>
-#include <seqan3/core/parallel/detail/reader_writer_manager.hpp>
+#include <seqan3/utility/parallel/detail/reader_writer_manager.hpp>
 #include <seqan3/utility/parallel/detail/spin_delay.hpp>
